pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apo-skillsmd"
version = "0.1.0"
description = "APO-SkillsMD research framework for evolving mixed-format agent skills."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "APO-SkillsMD Contributors" }
]
dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "pyyaml>=6.0",
  "python-frontmatter>=1.1",
  "bandit>=1.7",
  "gitpython>=3.1",
  "rank-bm25>=0.2",
  "numpy>=1.26",
  "rich>=13.7",
  "tenacity>=8.2",
  "diskcache>=5.6",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.52"]
openai = ["openai>=1.30"]
gemini = ["google-generativeai>=0.8"]
qwen = ["dashscope>=1.20"]
docker = ["docker>=7.0"]
embed = ["sentence-transformers>=2.7"]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
plugins = []

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
